requirements.txt
PySocks==1.6.8
dnspython==1.15.0